Add files via upload
Program created on 09.02.2024.
The idea for this program is what inspired me to start the 100 Days of Code course. I originally noted my training sessions manually in Notepad, but I thought there must be a better way. I wanted to be able to write a program to automate data entry. I created this program when I learned about saving files.